refactor: rename executables
